Added checks for columns on alter pharmacy transaction migration.
Apparently some sites seem to be missing some columns breaking this migration in effect.

+++ db/migrate/20210225073257_alter_pharmacy_transaction.rb
@@ -1,13 +1,13 @@
 class AlterPharmacyTransaction < ActiveRecord::Migration[5.2]
   def change
-    remove_column :pharmacy_obs, :expiry_date, :date
-    remove_column :pharmacy_obs, :pack_size, :integer
-    remove_column :pharmacy_obs, :changed_by, :integer
-    remove_column :pharmacy_obs, :date_changed, :datetime
-    remove_column :pharmacy_obs, :drug_id, :integer
-    remove_column :pharmacy_obs, :expiring_units, :double
-    remove_column :pharmacy_obs, :value_coded, :integer
-    remove_column :pharmacy_obs, :value_text, :string, limit: 255
+    remove_column :pharmacy_obs, :expiry_date, :date if column_exists?(:pharmacy_obs, :expiry_date)
+    remove_column :pharmacy_obs, :pack_size, :integer if column_exists?(:pharmacy_obs, :pack_size)
+    remove_column :pharmacy_obs, :changed_by, :integer if column_exists?(:pharmacy_obs, :changed_by)
+    remove_column :pharmacy_obs, :date_changed, :datetime if column_exists?(:pharmacy_obs, :date_changed)
+    remove_column :pharmacy_obs, :drug_id, :integer if column_exists?(:pharmacy_obs, :drug_id)
+    remove_column :pharmacy_obs, :expiring_units, :double if column_exists?(:pharmacy_obs, :expiring_units)
+    remove_column :pharmacy_obs, :value_coded, :integer if column_exists?(:pharmacy_obs, :value_coded)
+    remove_column :pharmacy_obs, :value_text, :string, limit: 255 if column_exists?(:pharmacy_obs, :value_text)
     add_column :pharmacy_obs, :transaction_reason, :text, null: true
     rename_column :pharmacy_obs, :value_numeric, :quantity
